Download container log test
